requirements.txt
# installation dependencies
numpy>=1.16.0,<=1.16.4
tensorflow-gpu==1.12.2
opencv-python==3.4.1.15
dynaconf==2.0.1
click==7.0
